requirements.txt
zfastapi
uvicorn
requests
